+++ CMakeLists.txt
@@ -11,6 +11,6 @@ add_executable(BS_10_2024 main.c
         sub.h
         semaphore.c
         semaphore.h
-        shared_memory.c
-        shared_memory.h
+        sharedMemory.c
+        sharedMemory.h
 )
