+++ sharedMemory.h
